refactor(import): change type  `Todo` for correct location file
